backend/pyproject.toml
[project]
name = "health-project-backend"
version = "2.0.0"
description = "AI-CHRMS: Centralized Health Record Management System API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.1",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    # Async task queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    # AI chatbot
    "openai>=1.0.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Field-level encryption
    "cryptography>=41.0.0",
]
